# docs/api-spec.yaml
openapi: 3.1.0
info:
  title: Movie Review Application API
  version: 1.0.0
  description: |
    RESTful API for the Movie Review Application.
    All state-changing endpoints require JWT authentication via httpOnly cookie.
    Admin-only endpoints additionally require the `isAdmin` claim in the JWT payload.

    **Authentication:** JWT access token is delivered as an httpOnly cookie named `access_token`.
    The frontend never accesses the token value directly. On 401 responses, the client
    should POST /auth/refresh to obtain a new access token.

    **CSRF Protection:** All state-changing requests (POST, PUT, PATCH, DELETE) must include
    the `X-CSRF-Token` header with the value read from the non-httpOnly `csrf_token` cookie.

    **Rate Limits:** Limit headers are returned on all responses:
    - `X-RateLimit-Limit`: request limit for the current window
    - `X-RateLimit-Remaining`: remaining requests
    - `X-RateLimit-Reset`: Unix timestamp when the window resets

    **Pagination:** List endpoints return a standard envelope:
    ```json
    { "data": [...], "meta": { "page": 1, "limit": 20, "total": 150, "totalPages": 8 } }
    ```

  contact:
    email: rawaldinesh13@dsu.ac.kr
  license:
    name: MIT
servers:
  - url: /api
    description: Application backend (proxied via Nginx at /api)

# ─────────────────────────────────────────────────────────────────────────────
# SECURITY SCHEMES
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token
      description: |
        JWT access token stored in httpOnly cookie. Set automatically by the browser
        after successful OAuth login. Expires in 15 minutes. Use POST /auth/refresh
        to renew silently.

  # ─────────────────────────────────────────────────────────────────────────
  # REUSABLE SCHEMAS
  # ─────────────────────────────────────────────────────────────────────────
  schemas:
...
    PaginationMeta:
      type: object
      required: [page, limit, total, totalPages]
      properties:
        page:
          type: integer
          minimum: 1
          example: 1
        limit:
          type: integer
          minimum: 1
          maximum: 100
          example: 20
        total:
          type: integer
          minimum: 0
          example: 150
        totalPages:
          type: integer
          minimum: 0
          example: 8

    # ── Error ────────────────────────────────────────────────────────────────
    ErrorResponse:
      type: object
      required: [statusCode, message]
      properties:
        statusCode:
          type: integer
          example: 400
        message:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
          example: "Validation failed"
        error:
          type: string
          example: "Bad Request"
        requestId:
          type: string
          format: uuid
          description: Correlation ID for log tracing

    # ── User ────────────────────────────────────────────────────────────────
    UserPublic:
      type: object
      description: Public user profile (no PII beyond display name and avatar)
      required: [id, displayName, avatarUrl, createdAt]
      properties:
        id:
          type: string
          format: uuid
        displayName:
          type: string
          example: "Dinesh Rawal"
        avatarUrl:
          type: string
          format: uri
          nullable: true
          example: "https://lh3.googleusercontent.com/a/..."
        createdAt:
          type: string
          format: date-time
          example: "2026-05-23T10:00:00Z"
    UserMe:
      allOf:
        - $ref: '#/components/schemas/UserPublic'
        - type: object
          required: [email, isAdmin]
          properties:
            email:
              type: string
              format: email
              description: Not displayed in public UI
              example: "rawaldinesh13@dsu.ac.kr"
            isAdmin:
              type: boolean
              example: false

    # ── Genre ────────────────────────────────────────────────────────────────
    Genre:
      type: object
      required: [id, name, tmdbGenreId]
      properties:
        id:
          type: integer
          example: 28
        name:
          type: string
          example: "Action"
        tmdbGenreId:
          type: integer
          example: 28

    # ── Movie ────────────────────────────────────────────────────────────────
    CastMember:
      type: object
      required: [name, character]
      properties:
        name:
          type: string
          example: "Brad Pitt"
        character:
          type: string
          example: "Tyler Durden"
        profilePath:
          type: string
          nullable: true
          description: TMDB profile image path; prepend https://image.tmdb.org/t/p/w185
          example: "/cckcYc2v0yh1tc9QjRelptcOBko.jpg"
    MovieCard:
      type: object
      description: Compact movie representation for list/search views
      required: [id, tmdbId, title, posterUrl, releaseDate, avgRating, ratingCount]
      properties:
        id:
          type: string
          format: uuid
        tmdbId:
          type: integer
          example: 550
        title:
          type: string
          example: "Fight Club"
        posterUrl:
          type: string
          format: uri
          nullable: true
          example: "https://image.tmdb.org/t/p/w500/pB8BM7pdSp6B6Ih7QZ4DrQ3PmJK.jpg"
        releaseDate:
          type: string
          format: date
          nullable: true
          example: "1999-10-15"
        releaseYear:
          type: integer
          nullable: true
          example: 1999
        avgRating:
          type: number
          format: float
          minimum: 0
          maximum: 5
          example: 4.3
        ratingCount:
          type: integer
          minimum: 0
          example: 247
        genres:
          type: array
          items:
            $ref: '#/components/schemas/Genre'
    MovieDetail:
      allOf:
        - $ref: '#/components/schemas/MovieCard'
        - type: object
          required: [backdropUrl, overview, runtime, originalLanguage, director, cast, cachedAt]
          properties:
            backdropUrl:
              type: string
              format: uri
              nullable: true
              example: "https://image.tmdb.org/t/p/original/hZkgoQYus5vegHoetLkCJzb17zJ.jpg"
            overview:
              type: string
              nullable: true
              example: "An insomniac office worker and a devil-may-care soap maker..."
            runtime:
              type: integer
              nullable: true
              description: Runtime in minutes
              example: 139
            originalLanguage:
              type: string
              nullable: true
              example: "en"
            trailerUrl:
              type: string
              format: uri
              nullable: true
              example: "https://www.youtube.com/watch?v=SUXWAEX2jlg"
            director:
              type: string
              nullable: true
              example: "David Fincher"
            cast:
              type: array
              maxItems: 10
              items:
                $ref: '#/components/schemas/CastMember'
            editorialNote:
              type: string
              nullable: true
              description: Admin-set editorial note (FR-083)
            cachedAt:
              type: string
              format: date-time
              description: When TMDB metadata was last refreshed
    MovieCreateRequest:
      type: object
      required: [tmdbId]
      properties:
        tmdbId:
          type: integer
          description: TMDB movie ID to import
          example: 550
    MovieUpdateRequest:
      type: object
      description: Admin-only fields that can be overridden locally
      properties:
        editorialNote:
          type: string
          nullable: true
          maxLength: 2000
          example: "A masterpiece of psychological thriller filmmaking."

    # ── Review ───────────────────────────────────────────────────────────────
    Review:
      type: object
      required: [id, user, movieId, body, isHidden, flagCount, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        user:
          $ref: '#/components/schemas/UserPublic'
        movieId:
          type: string
          format: uuid
        body:
          type: string
          maxLength: 500
          example: "A brutal, poetic film that stays with you long after the credits roll."
        isHidden:
          type: boolean
          description: If true, review is hidden (admin action); only visible to admins
          example: false
        flagCount:
          type: integer
          minimum: 0
          example: 0
        isEdited:
          type: boolean
          description: True if updatedAt > createdAt
          example: false
        createdAt:
          type: string
          format: date-time
          example: "2026-05-23T12:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2026-05-23T12:00:00Z"
    ReviewCreateRequest:
      type: object
      required: [body]
      properties:
        body:
          type: string
          minLength: 1
          maxLength: 500
          description: Review text. Must not be blank or whitespace-only.
          example: "A brutal, poetic film that stays with you long after the credits roll."
    ReviewUpdateRequest:
      type: object
      required: [body]
      properties:
        body:
          type: string
          minLength: 1
          maxLength: 500
          example: "Updated review text here."
...
    UserRoleUpdateRequest:
      type: object
      required: [isAdmin]
      properties:
        isAdmin:
          type: boolean
          description: true to promote to admin; false to revoke admin status
          example: true
...
  parameters:
    movieId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Internal movie UUID
    reviewId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Internal review UUID
    userId:
      name: userId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Internal user UUID
    pageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number (1-based)
    limitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Items per page

  # ─────────────────────────────────────────────────────────────────────────
  # REUSABLE RESPONSES
  # ─────────────────────────────────────────────────────────────────────────
  responses:
    Unauthorized:
      description: Missing or invalid JWT access token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            statusCode: 401
            message: "Unauthorized"
            error: "Unauthorized"
    Forbidden:
      description: Authenticated but insufficient role (not admin)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            statusCode: 403
            message: "Forbidden resource"
            error: "Forbidden"
    NotFound:
      description: Requested resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            statusCode: 404
            message: "Resource not found"
            error: "Not Found"
    UnprocessableEntity:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            statusCode: 422
            message: ["body must not be empty", "body must be shorter than or equal to 500 characters"]
            error: "Unprocessable Entity"
    TooManyRequests:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          schema:
            type: integer
        X-RateLimit-Remaining:
          schema:
            type: integer
        X-RateLimit-Reset:
          schema:
            type: integer
        Retry-After:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            statusCode: 429
            message: "ThrottlerException: Too Many Requests"
    Conflict:
      description: Resource already exists (e.g., duplicate review)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            statusCode: 409
            message: "You have already submitted a review for this movie"
            error: "Conflict"

# ─────────────────────────────────────────────────────────────────────────────
# PATHS
# ─────────────────────────────────────────────────────────────────────────────
paths:
...
  /movies:
    get:
      tags: [Movies]
      summary: List and search movies
      description: |
        Returns a paginated list of movies from the local cache (PostgreSQL).
        If `q` is provided, performs trigram search on movie titles.
        If no local movies match a search query, falls back to a TMDB API search
        and caches the results locally.
        TMDB "Popular" movies are returned when no query/filters are applied (page 1).
      operationId: listMovies
      security: []
      parameters:
        - name: q
          in: query
          schema:
            type: string
            maxLength: 200
          description: Search query (title search)
          example: "batman"
        - name: genre
          in: query
          schema:
            type: integer
          description: Filter by TMDB genre ID
          example: 28
        - name: yearFrom
          in: query
          schema:
            type: integer
            minimum: 1888
            maximum: 2100
          description: Filter movies released from this year (inclusive)
          example: 2000
        - name: yearTo
          in: query
          schema:
            type: integer
            minimum: 1888
            maximum: 2100
          description: Filter movies released up to this year (inclusive)
          example: 2020
        - name: sort
          in: query
          schema:
            type: string
            enum: [popularity, rating, year_desc, year_asc, title_asc]
            default: popularity
          description: Sort order
        - $ref: '#/components/parameters/pageParam'
        - $ref: '#/components/parameters/limitParam'
      responses:
        '200':
          description: Movie list
          headers:
            X-RateLimit-Limit:
              schema: { type: integer }
            X-RateLimit-Remaining:
              schema: { type: integer }
            X-RateLimit-Reset:
              schema: { type: integer }
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/MovieCard'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
                  tmdbFallback:
                    type: boolean
                    description: True if results came from TMDB API rather than local cache
                    example: false

    post:
      tags: [Movies]
      summary: Add movie from TMDB (admin)
      description: |
        Fetches movie metadata from TMDB by `tmdbId` and stores it in the local catalogue.
        Returns 409 if the movie is already in the catalogue.
        Admin only (FR-080).
      operationId: createMovie
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MovieCreateRequest'
      responses:
        '201':
          description: Movie added to catalogue
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MovieDetail'
        '400':
          description: Invalid TMDB ID or TMDB returned no result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /movies/{id}:
    parameters:
      - $ref: '#/components/parameters/movieId'

    get:
      tags: [Movies]
      summary: Get movie detail
      description: |
        Returns full movie detail including TMDB metadata, aggregate community rating,
        and genre list. Reviews are fetched separately via GET /movies/{id}/reviews.
      operationId: getMovie
      security: []
      responses:
        '200':
          description: Movie detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MovieDetail'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Movies]
      summary: Update movie (admin)
      description: |
        Allows admin to set an editorial note or override locally editable fields.
        TMDB metadata fields (title, poster, etc.) are NOT updated here — use POST /movies/{id}/sync.
        Admin only (FR-081, FR-083).
      operationId: updateMovie
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MovieUpdateRequest'
      responses:
        '200':
          description: Movie updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MovieDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Movies]
      summary: Delete movie from catalogue (admin)
      description: |
        Permanently removes the movie and all associated reviews and ratings (cascade).
        Admin must confirm intent via the `X-Confirm-Delete: true` header.
        Admin only (FR-082).
      operationId: deleteMovie
      security:
        - cookieAuth: []
      parameters:
        - name: X-Confirm-Delete
          in: header
          required: true
          schema:
            type: string
            enum: ["true"]
          description: Safety header — must be set to "true" to confirm destructive delete
      responses:
        '204':
          description: Movie deleted successfully
        '400':
          description: Missing confirmation header
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /movies/{id}/sync:
    parameters:
      - $ref: '#/components/parameters/movieId'

    post:
      tags: [Movies]
      summary: Trigger TMDB metadata refresh (admin)
      description: |
        Fetches the latest metadata from TMDB for this movie and updates the local record.
        Updates `cached_at`. Admin only (FR-081).
      operationId: syncMovie
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Metadata refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MovieDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '502':
          description: TMDB API unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                statusCode: 502
                message: "TMDB API is currently unavailable. Stale data retained."

  # ─── REVIEWS ────────────────────────────────────────────────────────────────
  /movies/{id}/reviews:
    parameters:
      - $ref: '#/components/parameters/movieId'

    get:
      tags: [Reviews]
      summary: List reviews for a movie
      description: |
        Returns all publicly visible reviews (is_hidden=false, deleted_at=null) for a movie,
        sorted newest-first. Paginated.
        Admins also see hidden reviews (indicated by isHidden=true in the response).
      operationId: listReviews
      security: []
      parameters:
        - $ref: '#/components/parameters/pageParam'
        - $ref: '#/components/parameters/limitParam'
      responses:
        '200':
          description: Review list
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Review'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
        '404':
          $ref: '#/components/responses/NotFound'

    post:
      tags: [Reviews]
      summary: Create review for a movie
      description: |
        Submits a new text review for a movie. One review per user per movie.
        Returns 409 if the user has already submitted a review.
        Rate limited: 20 requests/hour/user.
        Requires authentication (FR-050).
      operationId: createReview
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewCreateRequest'
      responses:
        '201':
          description: Review created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /reviews/{id}:
    parameters:
      - $ref: '#/components/parameters/reviewId'

    put:
      tags: [Reviews]
      summary: Edit own review
      description: |
        Updates the body of the authenticated user's review.
        Only the review author can edit their own review.
        Returns 403 if the requesting user is not the review author.
        `updatedAt` is set to now(); the edit timestamp is displayed in the UI.
        Requires authentication (FR-053).
      operationId: updateReview
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewUpdateRequest'
      responses:
        '200':
          description: Review updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'

    delete:
      tags: [Reviews]
      summary: Delete review (own or admin)
      description: |
        Permanently deletes a review.
        Review authors can delete their own reviews (FR-054).
        Admins can delete any review (FR-071).
        Admin deletions are recorded in the audit log.
        Returns 403 if the requester is neither the author nor an admin.
      operationId: deleteReview
      security:
        - cookieAuth: []
      responses:
        '204':
          description: Review deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /admin/reviews:
    get:
      tags: [Admin]
      summary: Admin moderation queue
      description: |
        Returns all reviews (including hidden) in the moderation queue.
        Filterable by status and sortable. Admin only (FR-070).
      operationId: adminListReviews
      security:
        - cookieAuth: []
      parameters:
        - name: filter
          in: query
          schema:
            type: string
            enum: [all, flagged, hidden, recent]
            default: all
          description: Filter moderation queue
        - name: sort
          in: query
          schema:
            type: string
            enum: [newest, most_flagged]
            default: newest
        - $ref: '#/components/parameters/pageParam'
        - $ref: '#/components/parameters/limitParam'
      responses:
        '200':
          description: Moderation queue
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Review'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/movies:
    get:
      tags: [Admin]
      summary: Admin movie management list
      description: Returns all movies in the local catalogue with management metadata (cached_at, ratingCount). Admin only.
      operationId: adminListMovies
      security:
        - cookieAuth: []
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: stale
          in: query
          schema:
            type: boolean
          description: Filter to only movies with stale TMDB data (cached_at older than 24h)
        - $ref: '#/components/parameters/pageParam'
        - $ref: '#/components/parameters/limitParam'
      responses:
        '200':
          description: Movie management list
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/MovieDetail'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/users:
    get:
      tags: [Admin]
      summary: Admin user management list
      description: Returns all registered users with admin status. Admin only (FR-084).
      operationId: adminListUsers
      security:
        - cookieAuth: []
      parameters:
        - name: q
          in: query
          schema:
            type: string
          description: Search by display name or email
        - $ref: '#/components/parameters/pageParam'
        - $ref: '#/components/parameters/limitParam'
      responses:
        '200':
          description: User list
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserMe'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/users/{userId}/role:
    parameters:
      - $ref: '#/components/parameters/userId'

    patch:
      tags: [Admin]
      summary: Promote or demote user admin role
      description: |
        Sets the `isAdmin` flag on a user account.
        Cannot demote self if sole remaining admin (FR-085).
        Recorded in audit log.
        Admin only (FR-084).
      operationId: updateUserRole
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRoleUpdateRequest'
      responses:
        '200':
          description: Role updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserMe'
        '400':
          description: Cannot demote sole admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                statusCode: 400
                message: "Cannot revoke admin status: you are the only administrator"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...